Localization [developed in the month of oct 2022 to december 2022]
I recently worked on a project to automate the process of checking localization status for ASINs on various marketplaces. Previously, this process was manual and time-consuming, requiring us to check the status individually on an internal web-based tool for each ASIN and marketplace. It took an average of 6 minutes per marketplace to upload the data and retrieve the final status report, totaling more than 60 minutes per run. This was causing delays in ASIN creation and localization Standard Operating Procedures (SOPs).

To address this issue, I divided the project into two parts: uploading the raw files and downloading the final dump. While the file uploading process was straightforward, the downloading part was more challenging and had previously been deemed not feasible by other automation experts. After a deep dive and exploration of all options, I came up with a solution. I created a directory in Outlook and set up a rule to move all localization emails to this directory. Then, I integrated the user's Outlook with a Python script to download all the localization emails and extract the final data as output.

As a result of this automation, the upload process now takes only 20 seconds for all 10 marketplaces combined, and the download process takes only 15 seconds. This has resulted in a 99.04% improvement in turnaround time and has saved a total head count of 3.68 annually.